add unit test job for emulator

+++ firmware/emulator/include/timer.h
@@ -1,30 +1,31 @@
 #pragma once
 
 #include <Arduino.h>
+#include <stdint.h>
 
 class Timer {
 public:
-    void setInterval(unsigned long interval_value);
-    unsigned long interval();
+    void setInterval(uint32_t interval_value);
+    uint32_t interval();
     bool isExpired(bool reset = false);
     void reset();
 
 protected:
-    unsigned long mInterval;
-    unsigned long mStart;
-    virtual unsigned long now() = 0; // Pure virtual function to get current time
+    uint32_t mInterval;
+    uint32_t mStart;
+    virtual uint32_t now() = 0; // Pure virtual function to get current time
 };
 
 class MillisTimer : public Timer {
 private:
-    unsigned long now() override;
+    uint32_t now() override;
 public:
     MillisTimer();
 };
 
 class MicrosTimer : public Timer {
 private:
-    unsigned long now() override;
+    uint32_t now() override;
 public:
     MicrosTimer();
 };
